Moving Files around to fix travis problem
this problem was caused by the last commit for the about and help menus under issue#66
