+++ src/main/java/nitezh/ministock/activities/menu/MenuScrollableAttributions.java
@@ -1,14 +1,14 @@
-package nitezh.ministock.activities.menu.help;
+package nitezh.ministock.activities.menu;
 
 import android.app.Activity;
 import android.os.Bundle;
 
 import nitezh.ministock.R;
 
-public class MenuScrollableEnteringStocks extends Activity {
+public class MenuScrollableAttributions extends Activity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.scroll_menu_entering_stocks);
+        setContentView(R.layout.scroll_menu_attributions);
     }
 }
