+++ src/main/java/nitezh/ministock/activities/menu/MenuScrollableEnteringStocks.java
@@ -1,14 +1,14 @@
-package nitezh.ministock.activities.menu.about;
+package nitezh.ministock.activities.menu;
 
 import android.app.Activity;
 import android.os.Bundle;
 
 import nitezh.ministock.R;
 
-public class MenuScrollableAttributions extends Activity {
+public class MenuScrollableEnteringStocks extends Activity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.scroll_menu_attributions);
+        setContentView(R.layout.scroll_menu_entering_stocks);
     }
 }
