+++ src/main/java/nitezh/ministock/activities/menu/MenuScrollableLicense.java
@@ -1,14 +1,14 @@
-package nitezh.ministock.activities.menu.about;
+package nitezh.ministock.activities.menu;
 
 import android.app.Activity;
 import android.os.Bundle;
 
 import nitezh.ministock.R;
 
-public class MenuScrollableTermsOfService extends Activity {
+public class MenuScrollableLicense extends Activity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.scroll_menu_terms_of_service);
+        setContentView(R.layout.scroll_menu_license);
     }
 }
