+++ src/main/java/nitezh/ministock/activities/menu/MenuScrollableOnlineHelp.java
@@ -1,14 +1,14 @@
-package nitezh.ministock.activities.menu.about;
+package nitezh.ministock.activities.menu;
 
 import android.app.Activity;
 import android.os.Bundle;
 
 import nitezh.ministock.R;
 
-public class MenuScrollableLicense extends Activity {
+public class MenuScrollableOnlineHelp extends Activity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.scroll_menu_license);
+        setContentView(R.layout.scroll_menu_online_help);
     }
 }
