+++ src/main/java/nitezh/ministock/activities/menu/MenuScrollableRecentChanges.java
@@ -1,14 +1,14 @@
-package nitezh.ministock.activities.menu.help;
+package nitezh.ministock.activities.menu;
 
 import android.app.Activity;
 import android.os.Bundle;
 
 import nitezh.ministock.R;
 
-public class MenuScrollableSelectingWidgetViews extends Activity {
+public class MenuScrollableRecentChanges extends Activity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.scroll_menu_selecting_widget_views);
+        setContentView(R.layout.scroll_menu_recent_changes);
     }
 }
