+++ src/main/java/nitezh/ministock/activities/menu/MenuScrollableSelectingWidgetViews.java
@@ -1,14 +1,14 @@
-package nitezh.ministock.activities.menu.help;
+package nitezh.ministock.activities.menu;
 
 import android.app.Activity;
 import android.os.Bundle;
 
 import nitezh.ministock.R;
 
-public class MenuScrollableUsingThePortfolio extends Activity {
+public class MenuScrollableSelectingWidgetViews extends Activity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.scroll_menu_using_the_portfolio);
+        setContentView(R.layout.scroll_menu_selecting_widget_views);
     }
 }
